organize components into their own files, add link functionality to events
