+++ apps/expo/src/app/home/index.tsx
@@ -30,10 +30,13 @@ import { useEffect, useState } from 'react';
 import { G, Path, Svg, Text as TextSVG } from 'react-native-svg';
 import { useColorScheme } from 'react-native';
 import { Toast, useToastController, useToastState } from '@tamagui/toast';
+import RestaurantTabs from '~/components/RestaurantTabs';
+import groupBy from '~/utils/groupBy';
 
 type Station = MenuWithRelations["stations"][0];
 type Dish = MenuWithRelations["stations"][0]["dishes"][0];
-type RestaurantName = NonNullable<ReturnType<typeof getRestaurantNameById>>;
+export type RestaurantName = NonNullable<ReturnType<typeof getRestaurantNameById>>;
+type PeriodName = NonNullable<ReturnType<typeof getCurrentPeriodName>>;
 
 // TODO: Replace with real user data
 const dummyUserPins = ["312"];
@@ -88,16 +91,17 @@ export function EventToast() {
         borderRadius="$20"
       >
         <Toast.Action altText='See Events' asChild>
-          <Button
-            backgroundColor={0}
-            pressTheme
-            size="$4"
-            onPress={() => toast.hide()}
-            circular
-            color="white"
-            icon={ArrowRight}
-            scaleIcon={1.5}
-          />
+          <Link href="/events/" asChild>
+            <Button
+              backgroundColor={0}
+              pressTheme
+              size="$4"
+              circular
+              color="white"
+              icon={ArrowRight}
+              scaleIcon={1.5}
+            />
+          </Link>
         </Toast.Action>
       </LinearGradient>
     </Toast>
@@ -105,8 +109,53 @@ export function EventToast() {
 }
 
 export function Home() {
-  const { selectedRestaurant } = useMenuStore();
+  const {
+    anteateryMenu,
+    brandywineMenu,
+    // setAnteateryMenu,
+    // setBrandywineMenu,
+  } = useMenuStore();
+
   const toast = useToastController();
+
+  const [date, setDate] = useState(new Date());
+  const [periodName, setPeriodName] = useState(getCurrentPeriodName());
+  const theme = useTheme();
+
+  // const anteateryMenu = anteateryData;
+  // const brandywineMenu = brandywineData;
+
+  // const [anteateryMenu, brandywineMenu] = api.useQueries((t) =>
+  //   (["anteatery", "brandywine"] as const).map((restaurantName) =>
+  //     t.menu.get({
+  //       date: date.toLocaleDateString("en-US"),
+  //       periodName,
+  //       restaurantName,
+  //     })));
+
+  // useEffect(() => {
+  //   if (anteateryMenu?.data) {
+  //     setAnteateryMenu(anteateryMenu.data);
+  //   }
+
+  //   if (brandywineMenu?.data) {
+  //     setBrandywineMenu(brandywineMenu.data);
+  //   }
+  // }, [anteateryMenu, brandywineMenu, setAnteateryMenu, setBrandywineMenu]);
+
+  // TODO: Could be better, maybe loading spinner
+
+  // if (anteateryMenu.isLoading || brandywineMenu.isLoading) {
+  //   return <View>Loading...</View>;
+  // }
+
+  // if (anteateryMenu.isError || brandywineMenu.isError || !anteateryMenu.data || !brandywineMenu.data) {
+  //   return <View>Error: {anteateryMenu.error ?? brandywineMenu.error}</View>;
+  // }
+
+  if (!anteateryMenu || !brandywineMenu) {
+    return <View>Loading...</View>;
+  }
 
   toast.show('There are 5 upcoming events.', {
     // message: 'See upcoming events',
@@ -118,26 +167,79 @@ export function Home() {
   });
 
   return (
-    <>
-      <Image
-        source={{
-          uri: selectedRestaurant === "brandywine" ?
-            "https://s3-media0.fl.yelpcdn.com/bphoto/P0DIhR8cO-JxYygc3V3aaQ/348s.jpg" :
-            "https://images.rsmdesign.com/7321bb55-579f-47fd-9f27-a6abf3e9826e.jpg"
-        }}
-        position='absolute'
-        zIndex={-1}
-        width={"100%"}
-        height={125}
-      />
-      <View height={65} />
-      <RestaurantTabs />
+    <RestaurantTabs>
       <EventToast />
-    </>
+      <XStack justifyContent='space-around'>
+        <PeriodPicker
+          periodName={periodName}
+          setPeriodName={setPeriodName}
+          color={theme.color?.val as string}
+        />
+        {/* TODO: Write a unit test for rendering and checking if onChange is triggered on event */}
+        <DateTimePicker
+          value={date}
+          mode="date"
+          onChange={(event: DateTimePickerEvent, selectedDate) => {
+            if (selectedDate) {
+              setDate(selectedDate);
+            }
+          }}
+        />
+      </XStack>
+
+      {[brandywineMenu, anteateryMenu].map((menu) => (
+        <Tabs.Content
+          key={menu.restaurantId}
+          value={getRestaurantNameById(menu.restaurantId as keyof typeof ID_TO_RESTAURANT)!}
+          alignItems="center"
+          flex={1}
+        >
+          <StationTabs stations={menu.stations} />
+        </Tabs.Content>
+      ))}
+    </RestaurantTabs>
   );
 };
 
-const CustomTab = ({ label }: { label: string }) => {
+interface PeriodPickerProps {
+  periodName: PeriodName;
+  setPeriodName: (periodName: PeriodName) => void;
+  color: string;
+}
+
+const PeriodPicker = ({
+  periodName,
+  setPeriodName,
+  color,
+}: PeriodPickerProps) => (
+  <Picker
+    style={{
+      width: 150,
+    }}
+    itemStyle={{
+      height: 50,
+      paddingVertical: 50,
+      fontSize: 18,
+      color,
+    }}
+    selectedValue={periodName}
+    onValueChange={(itemValue, _) =>
+      setPeriodName(itemValue)
+    }
+  >
+    {/* Create a Picker.Item for each period */}
+    {Object.entries(PERIOD_TO_ID).map(([period, id]) => (
+      <Picker.Item
+        key={id}
+        label={period}
+        value={id}
+      />
+    ))}
+  </Picker>
+);
+
+// Uses the svg from Figma
+export const TabSvg = ({ label }: { label: string }) => {
   const colorScheme = useColorScheme();
   const theme = useTheme();
   const deviceWidth = useWindowDimensions().width;
@@ -167,152 +269,21 @@ const CustomTab = ({ label }: { label: string }) => {
         >
           {label}
         </TextSVG>
-        {isCurrentlyClosed() && <TextSVG
+        <TextSVG
           x='50%'
           y='60%'
-          fill="firebrick"
+          fill={isCurrentlyClosed() ? "firebrick" : "forestgreen"}
           textAnchor='middle'
           alignmentBaseline="central"
           fontSize="18"
           fontWeight="bold"
         >
-          CLOSED
-        </TextSVG>}
+          {isCurrentlyClosed() ? "CLOSED" : "OPEN"}
+        </TextSVG>
       </G>
     </Svg>
   );
 };
-
-function RestaurantTabs() {
-  const {
-    selectedRestaurant,
-    anteateryMenu,
-    brandywineMenu,
-    setSelectedRestaurant,
-    // setAnteateryMenu,
-    // setBrandywineMenu,
-  } = useMenuStore();
-
-  // const anteateryMenu = anteateryData;
-  // const brandywineMenu = brandywineData;
-
-  const [date, setDate] = useState(new Date());
-  const [periodName, setPeriodName] = useState(getCurrentPeriodName());
-  // const [anteateryMenu, brandywineMenu] = api.useQueries((t) =>
-  //   (["anteatery", "brandywine"] as const).map((restaurantName) =>
-  //     t.menu.get({
-  //       date: date.toLocaleDateString("en-US"),
-  //       periodName,
-  //       restaurantName,
-  //     })));
-
-  // useEffect(() => {
-  //   if (anteateryMenu?.data) {
-  //     setAnteateryMenu(anteateryMenu.data);
-  //   }
-
-  //   if (brandywineMenu?.data) {
-  //     setBrandywineMenu(brandywineMenu.data);
-  //   }
-  // }, [anteateryMenu, brandywineMenu, setAnteateryMenu, setBrandywineMenu]);
-
-  const theme = useTheme();
-
-  // TODO: Could be better, maybe loading spinner
-  if (!anteateryMenu || !brandywineMenu) {
-    return <View>Loading...</View>;
-  }
-
-  // if (anteateryMenu.isLoading || brandywineMenu.isLoading) {
-  //   return <View>Loading...</View>;
-  // }
-
-  // if (anteateryMenu.isError || brandywineMenu.isError || !anteateryMenu.data || !brandywineMenu.data) {
-  //   return <View>Error: {anteateryMenu.error ?? brandywineMenu.error}</View>;
-  // }
-
-  const PeriodPicker = ({ color }: { color: string }) => (
-    <Picker
-      style={{
-        width: 150,
-      }}
-      itemStyle={{
-        height: 50,
-        paddingVertical: 50,
-        fontSize: 18,
-        color,
-      }}
-      selectedValue={periodName}
-      onValueChange={(itemValue, _) =>
-        setPeriodName(itemValue)
-      }
-    >
-      {/* Create a Picker.Item for each period */}
-      {Object.entries(PERIOD_TO_ID).map(([period, id]) => (
-        <Picker.Item
-          key={id}
-          label={period}
-          value={id}
-        />
-      ))}
-    </Picker>
-  );
-
-  return (
-    <Tabs
-      value={selectedRestaurant}
-      onValueChange={(value) => setSelectedRestaurant(value as RestaurantName)}
-      orientation="horizontal"
-      flexDirection="column"
-      width={"100%"}
-      height={"100%"}
-    >
-      <Tabs.List
-        borderRadius={"$20"}
-        // separator={<Separator vertical />}
-        // disablePassBorderRadius="bottom"
-        flexDirection='column'
-      >
-        <View width={"100%"} flexDirection='row'>
-          <Tabs.Tab flex={1} height={70} value="brandywine" opacity={selectedRestaurant === "brandywine" ? 1 : 0.5}>
-            <CustomTab label={"Brandywine"} />
-            {/* <H3 fontWeight={"800"}>Brandywine</H3> */}
-          </Tabs.Tab>
-          <Tabs.Tab flex={1} height={70} value="anteatery" opacity={selectedRestaurant === "anteatery" ? 1 : 0.5}>
-            <CustomTab label={"The Anteatery"} />
-            {/* <H3 fontWeight={"800"}>The Anteatery</H3> */}
-          </Tabs.Tab>
-        </View>
-      </Tabs.List>
-
-      <XStack justifyContent='space-around'>
-        <PeriodPicker color={theme.color?.val as string} />
-        {/* TODO: Write a unit test for rendering and checking if onChange is triggered on event */}
-        <DateTimePicker
-          value={date}
-          mode="date"
-          onChange={(event: DateTimePickerEvent, selectedDate) => {
-            if (selectedDate) {
-              setDate(selectedDate);
-            }
-          }}
-        />
-      </XStack>
-
-      {/* {[brandywineMenu.data, anteateryMenu.data].map((menu) => ( */}
-      {[brandywineMenu, anteateryMenu].map((menu) => (
-        <Tabs.Content
-          key={menu.restaurantId}
-          value={getRestaurantNameById(menu.restaurantId as keyof typeof ID_TO_RESTAURANT)!}
-          alignItems="center"
-          flex={1}
-        >
-          <StationTabs stations={menu.stations} />
-        </Tabs.Content>
-      ))}
-    </Tabs>
-  )
-}
 
 const StationTabs = ({ stations }: {
   stations: Station[],
@@ -372,25 +343,19 @@ const StationTabs = ({ stations }: {
   </Tabs>
 );
 
-const groupBy = <T, K extends keyof T>(arr: T[], key: (i: T) => K) =>
-  arr.reduce((groups, item) => {
-    (groups[key(item)] ||= []).push(item);
-    return groups;
-  }, {} as Record<K, T[]>);
-
 const Category = ({ category, dishes }: {
   category: string,
   dishes: Dish[],
 }) => (
   <YStack key={category} width={"100%"}>
-    <H3 fontWeight={"800"} marginTop="$5">{category}</H3>
+    <H3 fontWeight={"800"} marginTop="$5" paddingLeft="$2">{category}</H3>
     <YGroup bordered separator={<Separator borderWidth={1} />}>
       {dishes.map((dish) => (
         <DishCard key={dish.id} dish={dish} />
       ))}
     </YGroup>
   </YStack>
-)
+);
 
 const DishCard = ({ dish }: {
   dish: Dish,
@@ -404,30 +369,30 @@ const DishCard = ({ dish }: {
       }}
     >
       <ListItem pressTheme>
-        <XStack justifyContent='space-between'>
+        <XStack justifyContent='space-between' paddingRight="$4">
           <Image
             resizeMode="contain"
             alignSelf="center"
-            width={50}
-            height={50}
-            marginRight="$1"
+            width={65}
+            height={65}
+            marginRight="$3"
             source={{
               uri: 'https://images.rawpixel.com/image_png_1100/cHJpdmF0ZS9sci9pbWFnZXMvd2Vic2l0ZS8yMDIzLTExL2ZyZWVpbWFnZXNjb21wYW55X3Bob3RvX29mX2Nob2NvbGF0ZV9jaGlwX2Nvb2tpZV90b3Bfdmlld19pc29sYV8xOGVkY2ZiYS00ZTJjLTQ5MWItYjZiOC02ZGZjNmY1M2Y0OWIucG5n.png',
             }}
           />
-          <YStack gap="$1" width={"85%"} justifyContent='space-between' paddingVertical="$3" borderWidth={1} borderColor={"red"}>
+          <YStack gap="$1" width={"75%"} justifyContent='space-between' paddingTop="$4" paddingBottom="$3" >
             <XStack justifyContent='space-between'>
-              <Text fontWeight={"800"} fontSize={"$5"} borderWidth={1} borderColor={"red"}>{dish.name}</Text>
-              <Text textAlign='right' fontSize="$5" borderWidth={1} borderColor={"red"} fontWeight={"800"}>{dish.nutritionInfo.calories} cal</Text>
+              <Text fontWeight={"800"} fontSize={"$5"} >{dish.name}</Text>
+              <Text textAlign='right' fontSize="$5" fontWeight={"800"}>{dish.nutritionInfo.calories} cal</Text>
             </XStack>
-            <XStack justifyContent='space-between' borderWidth={1} borderColor={"red"}>
-              <XStack alignItems='center' gap="$1" borderWidth={1} borderColor={"red"} width={"70%"}>
+            <XStack justifyContent='space-between' >
+              <XStack alignItems='center' gap="$1" width={"70%"}>
                 <StarFull color="gold" scale={0.8} />
-                <Text><Text fontWeight="800" fontSize="$4">5.0</Text> <Text color={"gray"}>(10,000 reviews)</Text></Text>
+                <Text><Text fontWeight="800" fontSize="$4">5.0</Text> <Text color="gray">(10,000 reviews)</Text></Text>
               </XStack>
               {dummyUserPins.includes(dish.id) ?
-                <Button scale={0.8} fontWeight={"800"}>Unpin <PinOff /></Button> :
-                <Button scale={0.8} fontWeight={"800"}>Pin <Pin /></Button>
+                <Button minWidth={"48%"} scale={0.8} fontWeight={"800"}>Unpin <PinOff /></Button> :
+                <Button minWidth={"48%"} scale={0.8} fontWeight={"800"}>Pin <Pin /></Button>
               }
             </XStack>
           </YStack>
