+++ apps/expo/src/utils/groupBy.ts
@@ -1,0 +1,6 @@
+
+export default <T, K extends keyof T>(arr: T[], key: (i: T) => K) =>
+  arr.reduce((groups, item) => {
+    (groups[key(item)] ||= []).push(item);
+    return groups;
+  }, {} as Record<K, T[]>);
